TRUNK-4711: Do not suppress javadoc errors, increase error output limits
Now that all the warnings reported by the javadoc generation have
been fixed, remove the error suppression. Note that the this reveals
such problems in the javadocs that were not reported by compilation
when the suppression was active.

Increase the error output limit as by default, the compiler only outputs
the first 100 errors.